+++ protocols/cmp/keygen/round4.go
@@ -1,186 +1,161 @@
 package keygen
 
 import (
-	"bytes"
 	"errors"
-	"fmt"
 
-	"github.com/taurusgroup/cmp-ecdsa/internal/hash"
+	"github.com/taurusgroup/cmp-ecdsa/internal/proto"
 	"github.com/taurusgroup/cmp-ecdsa/internal/round"
-	"github.com/taurusgroup/cmp-ecdsa/pkg/paillier"
+	"github.com/taurusgroup/cmp-ecdsa/pkg/math/curve"
+	"github.com/taurusgroup/cmp-ecdsa/pkg/math/polynomial"
 	"github.com/taurusgroup/cmp-ecdsa/pkg/party"
-	"github.com/taurusgroup/cmp-ecdsa/pkg/pedersen"
 	"github.com/taurusgroup/cmp-ecdsa/pkg/protocol/message"
 	"github.com/taurusgroup/cmp-ecdsa/pkg/protocol/types"
 	zkmod "github.com/taurusgroup/cmp-ecdsa/pkg/zk/mod"
 	zkprm "github.com/taurusgroup/cmp-ecdsa/pkg/zk/prm"
-	zksch "github.com/taurusgroup/cmp-ecdsa/pkg/zk/sch"
 )
 
-var _ round.Round = (*round3)(nil)
+var _ round.Round = (*round4)(nil)
 
-type round3 struct {
-	*round2
-	// EchoHash = Hash(SSID, commitment₁, …, commitmentₙ)
-	EchoHash []byte
+type round4 struct {
+	*round3
 
-	// SchnorrCommitments[j] = Aⱼ
-	// Commitment for proof of knowledge in the last round
-	SchnorrCommitments map[party.ID]*zksch.Commitment // Aⱼ
+	// RID = ⊕ⱼ RIDⱼ
+	// Random ID generated by taking the XOR of all ridᵢ
+	RID RID
+	// ChainKey is a sequence of random bytes agreed upon together
+	ChainKey []byte
 }
 
 // ProcessMessage implements round.Round.
 //
-// - verify Hash(SSID, V₁, …, Vₙ) against received hash.
-// - store Fⱼ(X)
-//   - if keygen, verify Fⱼ(0) != ∞
-//   - if refresh, verify Fⱼ(0) == ∞
-// - verify length of Schnorr commitments
-// - validate Paillier
-// - validate Pedersen
-// - validate commitments.
-func (r *round3) ProcessMessage(j party.ID, content message.Content) error {
-	body := content.(*Keygen3)
-
-	if !bytes.Equal(body.HashEcho, r.EchoHash) {
-		return ErrRound3EchoHash
-	}
-
-	// Save all X, VSSCommitments
-	VSSPolynomial := body.VSSPolynomial
-	// check that the constant coefficient is 0
-	// if refresh then the polynomial is constant
-	if !(r.VSSSecret.Constant().IsZero() == VSSPolynomial.IsConstant) {
-		return ErrRound3VSSConstant
-	}
-	// check deg(Fⱼ) = t
-	if VSSPolynomial.Degree() != r.Threshold {
-		return ErrRound3VSSDegree
-	}
-
-	// Set Paillier
-	PaillierPublic, err := paillier.NewPublicKey(body.N)
+// - decrypt share
+// - verify VSS.
+func (r *round4) ProcessMessage(j party.ID, content message.Content) error {
+	body := content.(*Keygen4)
+	// decrypt share
+	DecryptedShare, err := r.PaillierSecret.Dec(body.Share)
 	if err != nil {
 		return err
 	}
-
-	// Verify Pedersen
-	Pedersen, err := pedersen.New(body.N, body.S, body.T)
-	if err != nil {
-		return err
+	Share := curve.NewScalarInt(DecryptedShare)
+	if DecryptedShare.Eq(Share.Int()) != 1 {
+		return ErrRound4Decrypt
 	}
 
-	// Verify decommit
-	if !r.HashForID(j).Decommit(r.Commitments[j], body.Decommitment,
-		body.RID, body.C, VSSPolynomial, body.SchnorrCommitments, Pedersen) {
-		return ErrRound3Decommit
+	// verify share with VSS
+	ExpectedPublicShare := r.VSSPolynomials[j].Evaluate(r.SelfID().Scalar()) // Fⱼ(i)
+	PublicShare := curve.NewIdentityPoint().ScalarBaseMult(Share)
+	// X == Fⱼ(i)
+	if !PublicShare.Equal(ExpectedPublicShare) {
+		return ErrRound4VSS
 	}
 
-	r.RIDs[j] = body.RID
-	r.ChainKeys[j] = body.RID
-	r.N[j] = body.N
-	r.S[j] = body.S
-	r.T[j] = body.T
-	r.PaillierPublic[j] = PaillierPublic
-	r.VSSPolynomials[j] = VSSPolynomial
-	r.SchnorrCommitments[j] = body.SchnorrCommitments
+	// verify zkmod
+	if !body.Mod.Verify(r.HashForID(j), zkmod.Public{N: r.N[j]}) {
+		return ErrRound4ZKMod
+	}
+
+	// verify zkprm
+	if !body.Prm.Verify(r.HashForID(j), zkprm.Public{N: r.N[j], S: r.S[j], T: r.T[j]}) {
+		return ErrRound4ZKPrm
+	}
+
+	r.ShareReceived[j] = Share
 	return nil
 }
 
 // Finalize implements round.Round
 //
-// - set rid = ⊕ⱼ ridⱼ and update hash state
-// - prove Nᵢ is Blum
-// - prove Pedersen parameters
-// - prove Schnorr for all coefficients of fᵢ(X)
-//   - if refresh skip constant coefficient
-//
-// - send proofs and encryption of share for Pⱼ.
-func (r *round3) Finalize(out chan<- *message.Message) (round.Round, error) {
-	// RID = ⊕ⱼ RIDⱼ
-	// c = ⊕ⱼ cⱼ
-	rid := newRID()
-	chainKey := r.PreviousChainKey
-	if chainKey == nil {
-		chainKeyRID := newRID()
-		for _, j := range r.PartyIDs() {
-			rid.XOR(r.RIDs[j])
-			chainKeyRID.XOR(r.ChainKeys[j])
-		}
-		chainKey = chainKeyRID
+// - sum of all received shares
+// - compute group public key and individual public keys
+// - recompute config SSID
+// - validate Config
+// - write new ssid hash to old hash state
+// - create proof of knowledge of secret.
+func (r *round4) Finalize(out chan<- *message.Message) (round.Round, error) {
+	// add all shares to our secret
+	UpdatedSecretECDSA := curve.NewScalar().Set(r.PreviousSecretECDSA)
+	for _, j := range r.PartyIDs() {
+		UpdatedSecretECDSA.Add(UpdatedSecretECDSA, r.ShareReceived[j])
 	}
-	// temporary hash which does not modify the state
+
+	// [F₁(X), …, Fₙ(X)]
+	ShamirPublicPolynomials := make([]*polynomial.Exponent, 0, len(r.VSSPolynomials))
+	for _, VSSPolynomial := range r.VSSPolynomials {
+		ShamirPublicPolynomials = append(ShamirPublicPolynomials, VSSPolynomial)
+	}
+
+	// ShamirPublicPolynomial = F(X) = ∑Fⱼ(X)
+	ShamirPublicPolynomial, err := polynomial.Sum(ShamirPublicPolynomials)
+	if err != nil {
+		return nil, err
+	}
+
+	// compute the new public key share Xⱼ = F(j) (+X'ⱼ if doing a refresh)
+	PublicData := make(map[party.ID]*Public, len(r.PartyIDs()))
+	for _, j := range r.PartyIDs() {
+		PublicECDSAShare := ShamirPublicPolynomial.Evaluate(j.Scalar())
+		PublicECDSAShare.Add(PublicECDSAShare, r.PreviousPublicSharesECDSA[j])
+		PublicData[j] = &Public{
+			ECDSA: PublicECDSAShare,
+			N:     r.N[j],
+			S:     r.S[j],
+			T:     r.T[j],
+		}
+	}
+
+	UpdatedConfig := &Config{
+		Threshold: uint32(r.Threshold),
+		Public:    PublicData,
+		RID:       r.RID.Copy(),
+		ChainKey:  r.ChainKey,
+		Secret: &Secret{
+			ID:    r.SelfID(),
+			ECDSA: UpdatedSecretECDSA,
+			P:     &proto.NatMarshaller{Nat: r.PaillierSecret.P()},
+			Q:     &proto.NatMarshaller{Nat: r.PaillierSecret.Q()},
+		},
+	}
+
+	// write new ssid to hash, to bind the Schnorr proof to this new config
+	// Write SSID, selfID to temporary hash
 	h := r.Hash()
-	_ = h.WriteAny(rid, r.SelfID())
+	_ = h.WriteAny(UpdatedConfig, r.SelfID())
 
-	// Prove N is a blum prime with zkmod
-	mod := zkmod.NewProof(h.Clone(), zkmod.Public{N: r.N[r.SelfID()]}, zkmod.Private{
-		P:   r.PaillierSecret.P(),
-		Q:   r.PaillierSecret.Q(),
-		Phi: r.PaillierSecret.Phi(),
-	})
+	proof := r.SchnorrRand.Prove(h, PublicData[r.SelfID()].ECDSA, UpdatedSecretECDSA)
 
-	// prove s, t are correct as aux parameters with zkprm
-	prm := zkprm.NewProof(h.Clone(), zkprm.Public{N: r.N[r.SelfID()], S: r.S[r.SelfID()], T: r.T[r.SelfID()]}, zkprm.Private{
-		Lambda: r.PedersenSecret,
-		Phi:    r.PaillierSecret.Phi(),
-	})
-
-	// create messages with encrypted shares
-	for _, j := range r.OtherPartyIDs() {
-		// compute fᵢ(j)
-		share := r.VSSSecret.Evaluate(j.Scalar())
-		// Encrypt share
-		C, _ := r.PaillierPublic[j].Enc(share.Int())
-
-		msg := r.MarshalMessage(&Keygen4{
-			Mod:   mod,
-			Prm:   prm,
-			Share: C,
-		}, j)
-		if err := r.SendMessage(msg, out); err != nil {
-			return r, err
-		}
+	// send to all
+	msg := r.MarshalMessage(&KeygenOutput{SchnorrResponse: proof}, r.OtherPartyIDs()...)
+	if err = r.SendMessage(msg, out); err != nil {
+		return r, err
 	}
 
-	// Write rid to the hash state
-	r.UpdateHashState(rid)
-	return &round4{
-		round3:   r,
-		RID:      rid,
-		ChainKey: chainKey,
+	r.UpdateHashState(UpdatedConfig)
+	return &output{
+		round4:        r,
+		UpdatedConfig: UpdatedConfig,
 	}, nil
 }
 
 // MessageContent implements round.Round.
-func (r *round3) MessageContent() message.Content { return &Keygen3{} }
+func (r *round4) MessageContent() message.Content { return &Keygen4{} }
 
 // Validate implements message.Content.
-func (m *Keygen3) Validate() error {
+func (m *Keygen4) Validate() error {
 	if m == nil {
-		return errors.New("keygen.round3: message is nil")
+		return errors.New("keygen.round4: message is nil")
 	}
-	if err := m.RID.Validate(); err != nil {
-		return fmt.Errorf("keygen.round3: %w", err)
+	if m.Mod == nil {
+		return errors.New("keygen.round4: zkmod proof is nil")
 	}
-
-	if err := m.Decommitment.Validate(); err != nil {
-		return fmt.Errorf("keygen.round3: %w", err)
+	if m.Prm == nil {
+		return errors.New("keygen.round4: zkprm proof is nil")
 	}
-
-	if m.N == nil || m.S == nil || m.T == nil {
-		return errors.New("keygen.round3: N,S,T invalid")
-	}
-
-	if m.VSSPolynomial == nil {
-		return errors.New("keygen.round3: VSSPolynomial is nil")
-	}
-
-	if l := len(m.HashEcho); l != hash.DigestLengthBytes {
-		return fmt.Errorf("keygen.round2: invalid echo hash length (got %d, expected %d)", l, hash.DigestLengthBytes)
+	if m.Share == nil {
+		return errors.New("keygen.round4: Share proof is nil")
 	}
 	return nil
 }
 
 // RoundNumber implements message.Content.
-func (m *Keygen3) RoundNumber() types.RoundNumber { return 4 }
+func (m *Keygen4) RoundNumber() types.RoundNumber { return 5 }
